requirement.txt
streamlit
langchain_google_genai
langchain_core
google-generativeai
python-dotenv
langchain 
langchain-community